Add `parent` project
This commit introduces the `parent` project as part of the BOM project.
This project used to live in `core` but has been moved to this project
instead, since it participates in releases independently of `core`.

As part of the process, the way the parent POM file has been generated
moves from a Groovy DSL based POM generator to a hybrid approach:

- a template pom.xml file is found in `src/pom-template.xml`
- properties are extracted into `libs.versions.toml` so that they can
be automatically upgraded by Renovatebot
- some properties are inlined from included BOMs (grpc, picocli)

A "maven-pom" plugin has been created so that if we want to generate
a new parent POM, say for libraries, it should be fairly straightforward.