Rename `bom` -> `platform`
The name "bom" was confusing with regards to the old BOM and the
new "core" BOM. We will now refer to the BOM which aggregates
several Micronaut modules (including core), driving releases, as
the "platform BOM".

Each module should have its own BOM, included in this project, and
this is where the "core BOM" is included.